♻️ [refactor] 검색 페이지 컴포넌트 분리 및 모듈 css 적용
검색 페이지 컴포넌트 분리 및 모듈 css 적용하였습니다.

fixes: 추후 리팩토링으로 변동 가능성 있음
