+++ src/components/search/Search.module.css
@@ -1,0 +1,73 @@
+.container {
+  @apply bg-black min-h-screen w-full flex flex-col items-center pl-8 pr-8 md:pl-2 md:pr-16 lg:pl-4 lg:pr-20;
+}
+.main {
+  @apply min-w-[75%] w-full  pt-10 md:pt-24 lg:pt-28 max-w-screen-xl mx-auto px-10 flex flex-col justify-center items-center;
+}
+
+.searchWrapper {
+  @apply mx-auto my-0  pt-[3.75em] pb-[0.25rem] px-0 border-b-[0.0625rem] border-b-white  min-w-[90%] max-w-full md:pt-8 lg:pt-12 flex justify-between items-center;
+}
+
+.searchInput {
+  @apply text-white border-neutral-600 box-border w-[calc(100%_-_3rem)] border-solid outline-none bg-transparent placeholder-gray-600 py-2 rounded-md text-3xl placeholder:text-3xl flex-grow;
+}
+
+.listWrapper {
+  @apply mt-[5rem] w-full px-[0.5rem] min-w-[95%] max-w-full flex justify-between ml-[5rem];
+}
+.recentViewWrapper {
+  @apply w-1/2 md:w-auto flex-grow flex flex-col basis-0;
+}
+.popularViewWrapper {
+  @apply w-1/2 md:w-auto h-full flex-grow border-l border-gray500 pl-[3.35rem] basis-0;
+}
+.searchResultWrapper {
+  @apply w-full mt-10;
+}
+.resultTitle {
+  @apply leading-[1.15] text-[1.375rem] font-semibold text-left text-white m-0 pt-0 pb-4 px-0 inline-block;
+}
+
+.resetButton {
+  @apply ml-[1.174rem] inline-flex pt-1;
+  @apply text-[1.175rem] font-normal text-[gray] leading-[normal];
+}
+.xIcon {
+  @apply w-[1.175rem] h-[1.175rem] inline-block bg-no-repeat bg-[50%_50%] bg-contain ml-[0.5rem] pt-[0.5rem];
+  @apply leading-normal;
+}
+
+.realtimeList {
+  @apply list-none m-2;
+  @apply text-lg text-gray-300 leading-relaxed;
+  @apply transition-colors duration-100;
+}
+.realtimeButton {
+  @apply bg-transparent cursor-pointer flex items-center;
+  @apply text-gray-400;
+  @apply transition-colors duration-100 ease-linear;
+}
+
+.recentTermButton {
+  @apply overflow-hidden  cursor-pointer;
+  @apply text-[#a6a6a6] text-[1.25rem] text-ellipsis whitespace-nowrap;
+}
+
+.recentXButton {
+  @apply min-w-[1rem] min-h-[1rem] bg-no-repeat bg-[50%_50%] opacity-70  cursor-pointer align-middle ml-4 pb-2;
+  @apply leading-[normal] text-base;
+}
+.swiperIncludeText {
+  @apply py-2 mt-1  opacity-70  overflow-hidden;
+  @apply text-[#dedede] text-base text-left whitespace-nowrap text-ellipsis;
+}
+.swiperList {
+  @apply opacity-70 mx-0 my-3 p-0;
+  @apply text-base text-[#dedede] leading-normal;
+  @apply transition-[color] duration-[0.1s];
+}
+.swiperContentsTitle {
+  @apply py-2 mt-2 overflow-hidden;
+  @apply text-[#dedede] opacity-70 text-base text-left whitespace-nowrap text-ellipsis;
+}
