+++ src/styles/index.css
@@ -101,7 +101,9 @@ html {
 .programNav .swiper-button-next::after,
 .programNav .swiper-button-prev::after,
 .categoryNav .swiper-button-next::after,
-.categoryNav .swiper-button-prev::after {
+.categoryNav .swiper-button-prev::after,
+.searchResult .swiper-button-next::after,
+.searchResult .swiper-button-prev::after {
   content: url(/assets/main-next.svg);
   background: url(/assets/main-next.svg);
   display: flex;
@@ -159,40 +161,37 @@ html {
 /* 내비게이션 관련 */
 .searchResult .swiper-button-next {
   position: absolute;
-  background: center / contain no-repeat
-      url(/assets/common/common-swiper-right-arrow.svg),
+  background: center / contain no-repeat,
     radial-gradient(
       106.25% 50% at 106.25% 50%,
       rgba(0, 0, 0, 0.9) 0%,
       rgba(0, 0, 0, 0.6) 52.6%,
       rgba(0, 0, 0, 0) 100%
     );
-  top: 45%;
+  top: 50%;
   transform: translateY(-50%);
   right: -0.5%;
   width: 3.5%;
   height: 100%;
-  opacity: 0.6;
   z-index: 10;
   transition: all 0.3s;
   cursor: pointer;
 }
 .searchResult .swiper-button-prev {
   position: absolute;
-  background: center / contain no-repeat
-      url(/assets/common/common-swiper-left-arrow.svg),
+  background: center / contain no-repeat,
     radial-gradient(
-      106.88% 50.3% at -6.88% 50%,
+      106.25% 50% at 106.25% 50%,
       rgba(0, 0, 0, 0.9) 0%,
-      rgba(0, 0, 0, 0.6) 53.21%,
+      rgba(0, 0, 0, 0.6) 52.6%,
       rgba(0, 0, 0, 0) 100%
     );
-  transform: translateY(-50%);
+  transform: translateY(-50%) scaleX(-1);
   left: -0.5%;
-  top: 45%;
+  top: 50%;
   width: 4%;
   height: 100%;
-  opacity: 0.6;
+
   z-index: 10;
   transition: all 0.3s;
   cursor: pointer;
